Improve `exportTrendPlots`
Fixed minor bugs in `exportTrendPlots` causing the XTickLabels to be inaccurate. Updated keyword properties for `boxplot` to make the boxplot panels look a little cleaner. The predicted marginal output (using fixed effects only from recovered GLME) is now superimposed on the boxplot per-day trends so that you can see what the model is fitting to the data.